backend/requirements.txt
Flask==2.3.2
PyPDF2==3.0.1
python-docx==0.8.11
spacy==3.5.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0.tar.gz
requests==2.28.1
beautifulsoup4==4.11.1
scikit-learn==1.2.2